Bring back Database.close (via java.io.Closeable)
Closing of a Database is frequently done in a synchronous context, it is
not a performance-critical operation, and the implementation for JDBC
cannot provide any benefits by wrapping it an extra thread.

Backend implementations are free to implement either `close` or
`shutdown` directly and provide the other one as a wrapper.