Allow classloader to be specified for classloading
Fixes #1152

Modifies the DatabaseConfig.forConfig and friends methods to take a
classloader to be used for classloading, as well as JdbcDataSource.
Unfortunately, JdbcBackend can't be modified without breaking binary
compatibility, since it's a trait.

Binary compatibility has been maintained, and I fixed the build so the
mima report works and is now run by travis.