Remove TypedNode:
TypedNodes always carried a type and didn’t lose it, even when
performing untyped transformations. Now that we have fully typed
transformations throughout the query compiler, this is not needed
anymore and `TypedNode` can be subsumed into `SimplyTypedNode`. A
TypedNode is now a SimplyTypedNode that infers a fixed type (specified
at construction time). This type is intrinsic to the node and will not
be lost when copying the node (just like any other constructor
parameter) but the real node type can be lost or reassigned normally.

- Add some extra `infer()` calls to instantiations of former TypedNodes
  throughout the query compiler.

- Add a special case for `NullaryNode` to `Node.mapChildren` to make
  `infer()` (and other calls) more efficient for `LiteralNode` (and
  other, less common NullaryNodes).