Remove AccessDriver
It’s been deprecated since 2.1, to be removed “when we drop support for
Java versions < 8”. This will probably not happen in 3.1 yet but
AccessDriver hasn’t been tested very well in 3.0 and would require
extra work to adapt it properly to the new compiler backend in 3.1.
Considering its very limited functionality and the buggy JDBC-ODBC
bridge, I doubt it will be missed by many people.