Database.forConfig uses provided config params `numThreads` and `queueSize`
The existing implementation of Database.forConfig was ignoring the `path`
passed to it when looking up the value of `numThreads` and `queueSize` in
the config.

I would have loved to provide a test together with the fix, but I would have
to resort to reflection to retrieve the values used to set up the thread pool,
and that seems fragile. Therefore, there is no test with this fix.

Conflicts:
	slick/src/main/scala/slick/jdbc/JdbcBackend.scala