More intelligent handling of String types on MySQL
Slick 3.0 changed the default text type for MySQL from `VARCHAR(254)` to
`TEXT` but this type is not allowed for primary keys or columns with
default values. We now check for the presence of `O.PrimaryKey` and
`O.Default` and revert back to `VARCHAR(254)` in these cases. This new
default is triggered by leaving the config key
`slick.driver.MySQL.defaultStringType` undefined.

To support this change, `JdbcTypesComponent.defaultSqlTypeName` and
`JdbcType.sqlTypeName` get access to the original `FieldSymbol` (when
used for column definitions as part of DDL generation) to look up any
column options they need.

Fixes #1129. Test in JdbcMiscTest.testNullability.