Prevent retry of lazy Connection acquisition attempt in BaseSession:
We used to mark the Session as open as soon as `conn` is accessed. If
the actual initialization (i.e. getting a JDBC Connection from the pool)
fails, an attempt to `close()` the Session would trigger another
initialization attempt. This change marks the Session as open only after
initialization has succeeded.

Fixes #1181.