Ensure non-empty “select” clauses in mergeToComprehensions:
Am empty select is filled with a dummy `LiteralNode(1)` to ensure
proper translation of such a Comprehension in `removeFieldNames` and
the code generator.

We also add an option to `RemoveFieldNames` for keeping StructNodes
intact when used in a “from” clause of a Comprehension, even when there
are no references. This can be used to force the creation of aliases
for dummy nodes in the code generator when required by a database.