Add `countDefined` operator:
Following on from #1239, which actually changed `.length` semantics for
all queries, not only those of a single column as documented, we add a
new `.countDefined` operator that performs a count of a single column,
omitting NULL rows.

As revealed by the new test cases, 3VL semantics for aggregation
operators were broken in QueryInterpreter. This is fixed, too.

Tests in AggregateTest. Fixes #1285.