Prevent repetition of bind variables used in GROUP BY:
Any expression from a GROUP BY clause may be inlined into later clauses
(like SELECT or ORDER BY) which can lead to duplication of expressions.
This is usually not a problem, in fact SQL is designed to allow reuse
of GROUP BY expressions without aggregation functions in these clauses.
This change prevents such reuse when a GROUP BY expression contains a
bind variable because the database could no longer know that the
expressions are indeed the same. If a leaked GROUP BY key is detected,
we push the source of the GroupBy into a subquery.

Test in AggregateTest.testFusedGroupBy. Fixes #1282.