Add Query#subquery as a precaution for implementing workarounds.
Bugs in the more aggressive query fusion in 3.1 or possible performance
issues in edge cases (as mentioned in #1282) should often be avoidable
by forcing a subquery to be created at a certain point. This change
adds an operator for this purpose to the API. We may remove it in a
later version when we feel confident that all problems of this kind have
been stamped out.

Test in NewQuerySemanticsTest.testNewFusion.