Run Reactive Streams tests with ScalaTest
The motivation for using sbt-testng-interface originally was to be
independent of Scala versions with a pure Java solution but this
back-fired because the actual interface implementation of
sbt-testng-interface is actually written in Scala. In order to build on
Scala 2.12 milestones, we are switching to ScalaTest for running the
TestNG tests. It has the same binary compatibility issues in theory but
there are builds available for Scala 2.12 milestone releases.