Fix 3VL handling in a scalar optimization
The case `(if(p) a else b) == v` in `optimizeScalar` was unnecessarily
general and getting it right under 3VL is hard (as witnessed by the fact
that the old implementation was broken). We now restrict it to
`(if(p) a else b).isNull`, the more specific case that is actually
needed to get rid of unnecessary code produced by Option rewriting in
earlier phases.

Test in JoinTest.testDiscriminatorCheck. Fixes #1300.