Select columns in the order in which the ResultConverters need them
Slick 3.1.0 abandoned the fixed result columns at the top level of a
query in order to remove unnecessary and duplicate columns. As a
side-effect the columns were randomly reordered. This turns out to be
a problem for `forceInsertQuery`, which expects the columns to be in
the specified order to match the insert operation’s column list. In
order to fix this problem `removeFieldNames` now uses a column order
which is derived from the ResultConverters.

Fixes #1338. Test in InsertTest.testForced.