Treat non-constant UUID defaults as no default.
If the database has a default which is not a constant value, Slick is not able
to represent the fact that the default for a column is a computed value. In
the case of UUIDs, this change will mean that if the UUID default is given as
an SQL function (for example, to generate a UUID), Slick will instead treat
the column as not having a default.