Ignore sqlite indices generated by autoindex feature.
This started breaking the tests when upgrading from org.xerial:sqlite-jdbc:3.8.7 to 3.8.11.2
But since they are automatically generated by sqlite it sort of makes sense not to put them into slick's
schema, so it doesn't try to create them. One might argue differently. We can change our minds in the
future if we need to.