Unified expansion of table types with same baseIdentity
In `expandTables` table fields used to be identified based on the
tables’s identity. This breaks when performing a union operation over
raw table rows where the right-hand side references more columns than
the left-hand side.

This situation can be prevented by taking field references of sibling
tables (with a different identity but the same baseIdentity) into
account when computing a tables’s structural type to ensure that the
structural types on both sides are compatible.

Test in UnionTest.testMappedUnion. Fixes #1571.