Prepare for 3.4.0-SNAPSHOT dev cycle (#1980)
The next release to be cut from `master` will be 3.4.0 during develop of 3.4.0 we check compatibility with 3.3.0.
The goal is not to stop any breaking change, but to make us aware. For each breaking change, we should add MiMa exclusions.

This will also help us decide when a PR can be backported in 3.3.x branch.  

When we cut 3.4.0 and create branch 3.4.x, we can remove the MiMa exclusions and start over again.