Replace sbt-version-scheme-enforcer with sbt-version-policy
sbt-version-scheme-enforcer is better IIUC when you're setting `version` explicitly

Instead, I'm setting the version automatically based on versionPolicyIntention and
an sbtdynver.GitDescribeOutput.

This obviates the need for `versionCheck`, which doesn't seem to support PVP anyway,
and means that after a PR the only thing that might need to be updated is
versionPolicyIntention.