Substitute null chars in postgres string type
\u0000 is not a valid character to submit to postgres. It will throw
if the char makes it this far.

Stripping them out of the supplied string is a very principled stance to make
and would be happy to have the PR turned down if it is deemed inappropriate.

To maintain backwards compatibility I have introduced the feature under a
flag people can optionally toggle in their postgres profile