Bring back MappedTo
Includes scala-2 and scala-3 versions of the macros. For scala-2,
the old one was brought back, but for scala-3, the api had to be
adjusted to comply with the new type projection rules. This difference
in API should not matter that much, as this is an implicit method.
The -source 3.0-migration option had to be removed as it caused the
added macro to not compile.